Move count_related() & dict_to_filter_params() to utilities.query
