+++ netbox/vpn/views.py
@@ -1,7 +1,7 @@
 from ipam.tables import RouteTargetTable
 from netbox.views import generic
 from tenancy.views import ObjectContactsView
-from utilities.utils import count_related
+from utilities.query import count_related
 from utilities.views import register_model_view
 from . import filtersets, forms, tables
 from .models import *
