+++ netbox/wireless/views.py
@@ -1,6 +1,6 @@
 from dcim.models import Interface
 from netbox.views import generic
-from utilities.utils import count_related
+from utilities.query import count_related
 from utilities.views import register_model_view
 from . import filtersets, forms, tables
 from .models import *
